speech_requirements.txt
torchaudio
ctc_segmentation
srt
praat-parselmouth
pangolinn
